Config service file with safe apps retrieval
