+++ src/routes/safe/components/Apps/components/AppsList.tsx
@@ -14,7 +14,7 @@ import { SAFELIST_ADDRESS } from 'src/routes/routes'
 import { useAppList } from '../hooks/useAppList'
 import { SAFE_APP_FETCH_STATUS, SafeApp } from '../types'
 import AddAppForm from './AddAppForm'
-import { AppData } from '../api/config-service'
+import { AppData } from 'src/logic/configService/index'
 
 const Wrapper = styled.div`
   height: 100%;
